Some fix and normalisation

+++ Makefile
@@ -10,7 +10,7 @@ DEST ?=$(BASEPATH)dist/
 CONTAINERSPATH ?=$(BASEPATH)containers/
 
 DEST_JS ?=$(DEST)js/
-SRC_JS ?=$(SRC)/js/
+SRC_JS ?=$(SRC)js/
 
 DEST_SASS ?=$(DEST)style/
 SRC_SASS ?=$(SRC)style/
@@ -18,8 +18,6 @@ SRC_SASS ?=$(SRC)style/
 CONTENT_PATH ?=content/
 LAYOUT_PATH ?=layouts/
 PARTIALS_PATH ?=partials/
-
-BEWATCH=./Makefile README.md ./package.json $(COMPILERS) $(SRC)*
 
 PACKAGE_NAME=$(shell cat "$(BASEPATH)package.json" | grep name | head -1 | awk -F: '{ print $2 }' | sed 's/[",]//g' | tr ':' '\n' | tail -1 | tr -d '[[:space:]]')
 
@@ -47,27 +45,31 @@ install-dev: .
 		curl -L -k https://github.com/emcrisostomo/fswatch/releases/download/${FSWV}/fswatch-${FSWV}.tar.gz | tar zx -C . && \
 		cd fswatch-${FSWV} && ./configure && make && make install
 
+
+BEWATCH:=./Makefile README.md ./package.json $(COMPILERS) $(SRC)*
 dev:
 		$(MAKE) browser-sync-start & \
 		fswatch -Ie "\.#.*" $(BEWATCH) | \
-		(while read file event; do \
-		ext=$${file##*.}; \
-		if [ $${ext} = "scss" ]; then \
-		$(MAKE) compile-sass; \
-		elif [ $${ext} = "js" ]; then \
-		$(MAKE) compile-js; \
-		elif [ $${ext} = "html" ]; then \
-		$(MAKE) compile-content; \
-		elif [ $${ext} = "md" ]; then \
-		$(MAKE) compile-content; \
-		else $(MAKE) build && $(MAKE) compile; \
-		fi \
-		done)
+		$(call WATCHER)
 
 browser-sync-start:
 		@if [ ! -z `which browser-sync` ]; \
 		then browser-sync start -s $(DEST) -f $(DEST) 1> /dev/null;\
 		else echo "We recommend to install browser-sync"; fi
+
+define WATCHER
+while read file event; do \
+		ext=$${file##*.}; \
+		if [ $${ext} = "scss" ]; then \
+		$(MAKE) compile-sass; \
+		elif [ $${ext} = "js" ]; then \
+		$(MAKE) compile-js; \
+		elif [ $${ext} = "html" || $${ext} = "md" ]; then \
+		$(MAKE) compile-content; \
+		else $(MAKE) build && $(MAKE) compile; \
+		fi \
+		done
+endef
 
 # END
 
@@ -104,49 +106,47 @@ clean-sass:
 
 
 # START --- Compiling utils
-
-compile: clean compile-js compile-sass compile-content ## Compile the website
+run := docker run -e "TYPE=tar" -i monera-${1}
 
 start-shared-volumes:
 		@$(eval MODULES:=$(shell docker run -d -t $(PACKAGE_NAME)-modules tail -f /dev/null))
 
-run-${CONTAINERS}:
-		docker run -i monera-${@}
+define RUN
+		docker run -e "TYPE=tar" -i monera-$1
+endef
 
-compile-js:
-		echo "\n\n--- Compiling JS"; \
-		$(MAKE) clean-js && \
-		cd $(SRC_JS) && \
-		tar c -h * | \
-		docker run -e "TYPE=tar" -i monera-es6 | \
-		docker run -e "TYPE=tar" -i monera-browserify | \
-		docker run -e "TYPE=tar" -i monera-uglify | \
-	  tar x -v -C "$(DEST_JS)" && \
-		echo "JS Compiled!"
+define RUNPKG
+		docker run -e "TYPE=tar" -i --volumes-from $(MODULES) monera-$1
+endef
 
-compile-cli: start-shared-volumes
-		cd src/cli && \
-		tar c -h * | \
-		docker run -e "TYPE=tar" -i monera-es6 | \
-		docker run -e "TYPE=tar" --volumes-from $(MODULES) -i monera-browserify --node | \
-	  tar x -v -C "$(DEST_JS)cli/" && \
-		docker kill $(MODULES)
-		echo "JS Compiled!"
+define COMPILE_START
+		echo "\n\n--- Compiling ${1}"; \
+		$(MAKE) clean-$1 && \
+		cd $2 && tar c -h $3
+endef
+
+define COMPILE_END
+		tar x -v -C "$(DEST)"
+		if[[ -n $(MODULES) ]]; then docker kill $(MODULES); fi
+		echo "--- Compiled!\n"
+endef
+
+compile: clean compile-js compile-sass compile-content ## Compile the website
+
+compile-js: start-shared-volumes
+		@$(call COMPILE_START,js,$(SRC_JS),*) | \
+		$(call RUN,es6) | $(call RUNPKG,browserify) | $(call RUN,uglify) | \
+		$(call COMPILE_END,js)
 
 compile-sass:
-		echo "\n\n--- Compiling Sass"; \
-		cd $(SRC_SASS) && \
-		tar c -h * | \
-		docker run -e "TYPE=tar" -i monera-sass | \
-	  tar x -v -C "$(DEST_SASS)" && \
-		echo "SASS Compiled!"
+		@$(call COMPILE_START,sass,$(SRC_SASS),*) | \
+		$(call RUN,sass) | \
+		$(call COMPILE_END,sass)
 
 compile-content:
-		echo "\n\n--- Compiling content"; \
-		cd $(SRC) && \
-		tar c -h $(CONTENT_PATH)* $(LAYOUT_PATH)* $(PARTIALS_PATH)* | docker run -i -e "TYPE=tar" monera-metalsmith | \
-	  tar x -v -C "$(DEST)" && \
-		echo "Content Compiled!"
+		@$(call COMPILE_START,content,$(SRC),"$(CONTENT_PATH)* $(LAYOUT_PATH)* $(PARTIALS_PATH)*") | \
+		$(call RUN,sass) | \
+		$(call COMPILE_END,content)
 
 # END
 
@@ -171,7 +171,7 @@ publish: build compile ## Publish
 
 # END
 
-.SILENT: clean-content clean-js clean-sass dev compile compile-js compile-sass compile-content
+#.SILENT: clean-content clean-js clean-sass dev compile compile-js compile-sass compile-content
 
 # START -- Utils
 bar := "\n--------------------------------------------------------\n"
@@ -179,7 +179,7 @@ title :="\n             M O N E R A   K I N G D O M"
 subtitle :="      containerised web development ready to use"
 more := " for more info visit: http://github.com/ideabile/monera"
 
-help:
+help: ## This help
 	@echo ${title}${bar}${subtitle}"\n"${more}${bar} && \
 	grep -E '^[a-zA-Z_-]+:.*?## .*$$' $(MAKEFILE_LIST) | sort | awk 'BEGIN {FS = ":.*?## "}; {printf "\033[36m%-30s\033[0m %s\n", $$1, $$2}' && \
 	echo ${bar}
